requirements.txt
langchain==0.3.21
langchain_core==0.3.49
langchain_community==0.3.20
langchainhub==0.1.21
langserve==0.3.1
langchain_openai==0.3.11
langchain-groq==0.3.1
langchain_cohere==0.4.3
python-dotenv
streamlit
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
wikipedia
arxiv
sentence_transformers
PyPDF2
pydantic
httpx==0.27.2
ollama